Add default user-agent + enable custom one
Every library appends their user-agent to the one that's been set at a level above them.

Hence any provided user-agent comes before the default user agent.

To provide context, the version number of `@fastify/oauth2` + the home page of it is included in the User-Agent. That way an API can diagnose possibly faulty consumers and file an issue.

If someone wouldn't want to expose this, they can always set a custom User-Agent header directly in the http settings. Especially if #225 gets merged.