Add default user-agent + enable custom one (#226)
* Fix `test:coverage` command

* Add default user-agent + enable custom one

Every library appends their user-agent to the one that's been set at a level above them.

Hence any provided user-agent comes before the default user agent.

To provide context, the version number of `@fastify/oauth2` + the home page of it is included in the User-Agent. That way an API can diagnose possibly faulty consumers and file an issue.

If someone wouldn't want to expose this, they can always set a custom User-Agent header directly in the http settings. Especially if #225 gets merged.

* Update index.js

Co-authored-by: Manuel Spigolon <behemoth89@gmail.com>

* Fix feedback, allow disabling of user-agent

* Simplify user-agent string

* Move `credentials` initialization to the top

* Naming tweak in test

* Remove multi-`user-agent` use

The RFC 9110 states:

"The User-Agent field value consists of one or more product identifiers, each followed by zero or more comments (Section 5.6.5), which together identify the user agent software and its significant subproducts. By convention, the product identifiers are listed in decreasing order of their significance for identifying the user agent software."

But since many are not aware of it, it may be better to remove it

---------

Co-authored-by: Manuel Spigolon <behemoth89@gmail.com>